add necessary changes to support CRD sync
