+++ incubator/virtualcluster/pkg/util/mccontroller/mccontroller.go
@@ -20,14 +20,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"k8s.io/client-go/rest"
 	"net/http"
 	"sync"
 	"time"
 
 	v1 "k8s.io/api/core/v1"
-	extensionsv1beta1 "k8s.io/api/extensions/v1beta1"
-	schedulingv1 "k8s.io/api/scheduling/v1"
-	storagev1 "k8s.io/api/storage/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -105,6 +103,7 @@ type ClusterInterface interface {
 	AddEventHandler(runtime.Object, clientgocache.ResourceEventHandler) error
 	GetClientSet() (clientset.Interface, error)
 	GetDelegatingClient() (client.Client, error)
+	GetRestConfig() *rest.Config
 	Cache
 }
 
@@ -529,74 +528,4 @@ func filterSuperClusterSchedulePod(c *MultiClusterController, req reconciler.Req
 	}
 
 	return cname != constants.SuperClusterID
-}
-
-func getTargetObject(objectType runtime.Object) runtime.Object {
-	switch objectType.(type) {
-	case *v1.ConfigMap:
-		return &v1.ConfigMap{}
-	case *v1.Namespace:
-		return &v1.Namespace{}
-	case *v1.Node:
-		return &v1.Node{}
-	case *v1.Event:
-		return &v1.Event{}
-	case *v1.Pod:
-		return &v1.Pod{}
-	case *v1.Secret:
-		return &v1.Secret{}
-	case *v1.Service:
-		return &v1.Service{}
-	case *v1.ServiceAccount:
-		return &v1.ServiceAccount{}
-	case *storagev1.StorageClass:
-		return &storagev1.StorageClass{}
-	case *v1.PersistentVolumeClaim:
-		return &v1.PersistentVolumeClaim{}
-	case *v1.PersistentVolume:
-		return &v1.PersistentVolume{}
-	case *v1.Endpoints:
-		return &v1.Endpoints{}
-	case *schedulingv1.PriorityClass:
-		return &schedulingv1.PriorityClass{}
-	case *extensionsv1beta1.Ingress:
-		return &extensionsv1beta1.Ingress{}
-	default:
-		return nil
-	}
-}
-
-func getTargetObjectList(objectType runtime.Object) runtime.Object {
-	switch objectType.(type) {
-	case *v1.ConfigMap:
-		return &v1.ConfigMapList{}
-	case *v1.Namespace:
-		return &v1.NamespaceList{}
-	case *v1.Node:
-		return &v1.NodeList{}
-	case *v1.Event:
-		return &v1.EventList{}
-	case *v1.Pod:
-		return &v1.PodList{}
-	case *v1.Secret:
-		return &v1.SecretList{}
-	case *v1.Service:
-		return &v1.ServiceList{}
-	case *v1.ServiceAccount:
-		return &v1.ServiceAccountList{}
-	case *storagev1.StorageClass:
-		return &storagev1.StorageClassList{}
-	case *v1.PersistentVolumeClaim:
-		return &v1.PersistentVolumeClaimList{}
-	case *v1.PersistentVolume:
-		return &v1.PersistentVolumeList{}
-	case *v1.Endpoints:
-		return &v1.EndpointsList{}
-	case *schedulingv1.PriorityClass:
-		return &schedulingv1.PriorityClassList{}
-	case *extensionsv1beta1.Ingress:
-		return &extensionsv1beta1.IngressList{}
-	default:
-		return nil
-	}
 }
